Refactor helper modules
