+++ lib/nimble_template/helpers/dependency.ex
@@ -1,0 +1,5 @@
+defmodule NimbleTemplate.DependencyHelper do
+  def fetch_and_install_dependencies(),
+    do:
+      if(Mix.shell().yes?("\nFetch and install dependencies?"), do: Mix.shell().cmd("mix deps.get"))
+end
